+++ src/Fable.Helpers.ReactNative.fs
@@ -26,6 +26,21 @@ module Props =
     | Center
     | [<CompiledName("flex-end")>] FlexEnd
     | Stretch
+    | Baseline
+
+    [<StringEnum; RequireQualifiedAccess>]
+    type ContentAlignment =
+    | [<CompiledName("flex-start")>] FlexStart
+    | Center
+    | [<CompiledName("flex-end")>] FlexEnd
+    | Stretch
+    | [<CompiledName("space-between")>] SpaceBetween
+    | [<CompiledName("space-around")>] SpaceAround
+
+    [<StringEnum; RequireQualifiedAccess>]
+    type Display =
+    | None
+    | Flex
 
     [<StringEnum; RequireQualifiedAccess>]
     type ItemAlignment =
@@ -33,7 +48,8 @@ module Props =
     | Center
     | [<CompiledName("flex-end")>] FlexEnd
     | Stretch
-
+    | Baseline
+    
     [<StringEnum; RequireQualifiedAccess>]
     type TextAlignment =
     | Auto
@@ -57,6 +73,7 @@ module Props =
     | [<CompiledName("flex-end")>] FlexEnd
     | [<CompiledName("space-between")>] SpaceBetween
     | [<CompiledName("space-around")>] SpaceAround
+    | [<CompiledName("space-evenly")>] SpaceEvenly
 
     [<StringEnum; RequireQualifiedAccess>]
     type FlexDirection =
@@ -67,9 +84,7 @@ module Props =
 
     [<StringEnum; RequireQualifiedAccess>]
     type FlexWrap =
-    | Row
-    | [<CompiledName("wrap")>] Wrap
-    | Column
+    | Wrap
     | [<CompiledName("nowrap")>] NoWrap
 
     [<StringEnum; RequireQualifiedAccess>]
@@ -89,11 +104,15 @@ module Props =
     | [<CompiledName("web-search")>] WebSearch
 
     [<StringEnum; RequireQualifiedAccess>]
+    type Overflow =
+    | Visible
+    | Hidden
+    | Scroll
+
+    [<StringEnum; RequireQualifiedAccess>]
     type Position =
-    | Row
-    | [<CompiledName("absolute")>] Absolute
-    | Column
-    | [<CompiledName("relative")>] Relative
+    | Absolute
+    | Relative
 
     [<StringEnum; RequireQualifiedAccess>]
     type ResizeMode =
@@ -138,10 +157,6 @@ module Props =
     | [<CompiledName("locked-open")>] LockedOpen
 
     [<StringEnum; RequireQualifiedAccess>]
-    type AlignSelf =
-        | Auto
-
-    [<StringEnum; RequireQualifiedAccess>]
     type TextDecorationStyle =
         | Solid | Double | Dotted | Dashed
 
@@ -173,9 +188,10 @@ module Props =
 
     [<StringEnum; RequireQualifiedAccess>]
     type TextDecorationLine =
-        | None | Underline
-        | ``Line-through``
-        | ``Underline line-through``
+        | None
+        | Underline
+        | [<CompiledName("line-through")>] LineThrough
+        | [<CompiledName("underline line-through")>] UnderlineLineThrough
 
     [<StringEnum; RequireQualifiedAccess>]
     type LineBreakMode =
@@ -194,7 +210,7 @@ module Props =
         | Solid | Dotted | Dashed
 
     [<StringEnum; RequireQualifiedAccess>]
-    type Overflow =
+    type ImageOverflow =
         | Visible | Hidden
 
     [<StringEnum; RequireQualifiedAccess>]
@@ -231,7 +247,18 @@ module Props =
 
     [<StringEnum; RequireQualifiedAccess>]
     type SystemIcon =
-        | Bookmarks | Contacts | Downloads | Favorites | Featured | History | More | ``Most-recent`` | ``Most-viewed`` | Recents | Search | ``Top-rated``
+        | Bookmarks
+        | Contacts
+        | Downloads
+        | Favorites
+        | Featured 
+        | History
+        | More
+        | [<CompiledName("most-recent")>] MostRecent
+        | [<CompiledName("most-viewed")>] MostViewed
+        | Recents
+        | Search
+        | [<CompiledName("top-rated")>] TopRated
 
     [<StringEnum; RequireQualifiedAccess>]
     type Dim =
@@ -264,6 +291,9 @@ module Props =
     [<StringEnum; RequireQualifiedAccess>]
     type Direction =
         | Horizontal | Vertical
+
+    type ISizeUnit =
+        interface end
 
     type IStyle =
         interface end
@@ -346,7 +376,7 @@ module Props =
     type IListViewProperties =
         interface end
 
-    type IFlatListProperties =
+    type IFlatListProperties<'a> =
         interface end
 
     type IScrollViewProperties =
@@ -424,66 +454,87 @@ module Props =
         interface ITransformsStyle
 
     type FlexStyle =
+        | AlignContent of ContentAlignment
         | AlignItems of ItemAlignment
         | AlignSelf of Alignment
         | AspectRatio of float
         | BorderBottomWidth of float
+        | BorderEndWidth of float
         | BorderLeftWidth of float
         | BorderRightWidth of float
+        | BorderStartWidth of float
         | BorderTopWidth of float
         | BorderWidth of float
-        | Bottom of float
+        | Bottom of ISizeUnit
+        | Display of Display
+        | End of ISizeUnit
         | Flex of float
+        | FlexBasis of ISizeUnit
         | FlexDirection of FlexDirection
+        | FlexGrow of float
+        | FlexShrink of float
         | FlexWrap of FlexWrap
-        | Height of float
+        | Height of ISizeUnit
         | JustifyContent of JustifyContent
-        | Left of float
-        | MinWidth of float
-        | MaxWidth of float
-        | MinHeight of float
-        | MaxHeight of float
-        | Margin of float
-        | MarginBottom of float
-        | MarginHorizontal of float
-        | MarginLeft of float
-        | MarginRight of float
-        | MarginTop of float
-        | MarginVertical of float
-        | Padding of float
-        | PaddingBottom of float
-        | PaddingHorizontal of float
-        | PaddingLeft of float
-        | PaddingRight of float
-        | PaddingTop of float
-        | PaddingVertical of float
+        | Left of ISizeUnit
+        | Margin of ISizeUnit
+        | MarginBottom of ISizeUnit
+        | MarginEnd of ISizeUnit
+        | MarginHorizontal of ISizeUnit
+        | MarginLeft of ISizeUnit
+        | MarginRight of ISizeUnit
+        | MarginStart of ISizeUnit
+        | MarginTop of ISizeUnit
+        | MarginVertical of ISizeUnit
+        | MaxHeight of ISizeUnit
+        | MaxWidth of ISizeUnit
+        | MinHeight of ISizeUnit
+        | MinWidth of ISizeUnit
+        | Overflow of Overflow
+        | Padding of ISizeUnit
+        | PaddingBottom of ISizeUnit
+        | PaddingEnd of ISizeUnit
+        | PaddingHorizontal of ISizeUnit
+        | PaddingLeft of ISizeUnit
+        | PaddingRight of ISizeUnit
+        | PaddingStart of ISizeUnit
+        | PaddingTop of ISizeUnit
+        | PaddingVertical of ISizeUnit
         | Position of Position
-        | Right of float
-        | Top of float
-        | Width of float
+        | Right of ISizeUnit
+        | Start of ISizeUnit
+        | Top of ISizeUnit
+        | Width of ISizeUnit
         | ZIndex of float
         interface IFlexStyle
 
 
     type ViewStyle =
-        | BackfaceVisibility of string
+        | BackfaceVisibility of BackfaceVisibility
         | BackgroundColor of string
         | BorderBottomColor of string
+        | BorderBottomEndRadius of float
         | BorderBottomLeftRadius of float
         | BorderBottomRightRadius of float
+        | BorderBottomStartRadius of float
         | BorderBottomWidth of float
         | BorderColor of string
+        | BorderEndColor of string
         | BorderLeftColor of string
+        | BorderLeftWidth of float
         | BorderRadius of float
         | BorderRightColor of string
         | BorderRightWidth of float
+        | BorderStartColor of string
         | BorderStyle of BorderStyle
         | BorderTopColor of string
+        | BorderTopEndRadius of float
         | BorderTopLeftRadius of float
         | BorderTopRightRadius of float
+        | BorderTopStartRadius of float
         | BorderTopWidth of float
+        | BorderWidth of float
         | Opacity of float
-        | Overflow of Overflow
         | ShadowColor of string
         | ShadowOffset of obj
         | ShadowOpacity of float
@@ -564,8 +615,8 @@ module Props =
         | AllowFontScaling of bool
         | LineBreakMode of LineBreakMode
         | NumberOfLines of float
-        | OnLayout of Func<LayoutChangeEvent, unit>
-        | OnPress of (unit->unit)
+        | OnLayout of (LayoutChangeEvent -> unit)
+        | OnPress of (unit -> unit)
         | Style of IStyle list
         | TestID of string
         interface ITextProperties
@@ -585,7 +636,7 @@ module Props =
             | ClearButtonMode of string
             | ClearTextOnFocus of bool
             | EnablesReturnKeyAutomatically of bool
-            | OnKeyPress of (unit->unit)
+            | OnKeyPress of (unit -> unit)
             | SelectionState of obj
             interface ITextInputIOSProperties
 
@@ -607,14 +658,14 @@ module Props =
             | KeyboardType of KeyboardType
             | MaxLength of float
             | Multiline of bool
-            | OnBlur of (unit->unit)
+            | OnBlur of (unit -> unit)
             | OnChange of (obj -> unit)
             | OnChangeText of (string -> unit)
             | OnEndEditing of (obj -> unit)
-            | OnFocus of (unit->unit)
-            | OnLayout of Func<obj, unit>
-            | OnSelectionChange of (unit->unit)
-            | OnSubmitEditing of Func<obj, unit>
+            | OnFocus of (unit -> unit)
+            | OnLayout of (obj -> unit)
+            | OnSelectionChange of (unit -> unit)
+            | OnSubmitEditing of (obj -> unit)
             | Password of bool
             | Placeholder of string
             | PlaceholderTextColor of string
@@ -634,7 +685,7 @@ module Props =
             | ContentInsetStart of float
             | Logo of obj
             | NavIcon of obj
-            | OnIconClicked of (unit->unit)
+            | OnIconClicked of (unit -> unit)
             | OverflowIcon of obj
             | Rtl of bool
             | Style of IStyle list
@@ -647,16 +698,16 @@ module Props =
             interface IToolbarAndroidProperties
 
     type GestureResponderHandlers =
-        | OnStartShouldSetResponder of Func<GestureResponderEvent, bool>
-        | OnMoveShouldSetResponder of Func<GestureResponderEvent, bool>
-        | OnResponderGrant of Func<GestureResponderEvent, unit>
-        | OnResponderReject of Func<GestureResponderEvent, unit>
-        | OnResponderMove of Func<GestureResponderEvent, unit>
-        | OnResponderRelease of Func<GestureResponderEvent, unit>
-        | OnResponderTerminationRequest of Func<GestureResponderEvent, bool>
-        | OnResponderTerminate of Func<GestureResponderEvent, unit>
-        | OnStartShouldSetResponderCapture of Func<GestureResponderEvent, bool>
-        | OnMoveShouldSetResponderCapture of (unit->unit)
+        | OnStartShouldSetResponder of (GestureResponderEvent -> bool)
+        | OnMoveShouldSetResponder of (GestureResponderEvent -> bool)
+        | OnResponderGrant of (GestureResponderEvent -> unit)
+        | OnResponderReject of (GestureResponderEvent -> unit)
+        | OnResponderMove of (GestureResponderEvent -> unit)
+        | OnResponderRelease of (GestureResponderEvent -> unit)
+        | OnResponderTerminationRequest of (GestureResponderEvent -> bool)
+        | OnResponderTerminate of (GestureResponderEvent -> unit)
+        | OnStartShouldSetResponderCapture of (GestureResponderEvent -> bool)
+        | OnMoveShouldSetResponderCapture of (GestureResponderEvent -> unit)
         interface IGestureResponderHandlers
 
     type ViewPropertiesIOS =
@@ -677,9 +728,9 @@ module Props =
         | AccessibilityLabel of string
         | Accessible of bool
         | HitSlop of obj
-        | OnAcccessibilityTap of (unit->unit)
-        | OnLayout of Func<LayoutChangeEvent, unit>
-        | OnMagicTap of (unit->unit)
+        | OnAcccessibilityTap of (unit -> unit)
+        | OnLayout of (LayoutChangeEvent -> unit)
+        | OnMagicTap of (unit -> unit)
         | PointerEvents of PointerEvents
         | RemoveClippedSubviews of bool
         | Style of IStyle list
@@ -689,9 +740,9 @@ module Props =
     type ViewPagerAndroidProperties =
         | InitialPage of int
         | ScrollEnabled of bool
-        | OnPageScroll of Func<NativeSyntheticEvent<ViewPagerAndroidOnPageScrollEventData>, unit>
-        | OnPageSelected of Func<NativeSyntheticEvent<ViewPagerAndroidOnPageSelectedEventData>, unit>
-        | OnPageScrollStateChanged of Func<ScrollState, unit>
+        | OnPageScroll of (NativeSyntheticEvent<ViewPagerAndroidOnPageScrollEventData> -> unit)
+        | OnPageSelected of (NativeSyntheticEvent<ViewPagerAndroidOnPageSelectedEventData> -> unit)
+        | OnPageScrollStateChanged of (ScrollState -> unit)
         | KeyboardDismissMode of KeyboardDismissMode
         | PageMargin of float
         | Style of IStyle list
@@ -713,7 +764,7 @@ module Props =
         | AllowsInlineMediaPlayback of bool
         | Bounces of bool
         | DecelerationRate of DecelerationRate
-        | OnShouldStartLoadWithRequest of Func<WebViewIOSLoadRequestEvent, bool>
+        | OnShouldStartLoadWithRequest of (WebViewIOSLoadRequestEvent -> bool)
         | ScrollEnabled of bool
         interface IWebViewPropertiesIOS
 
@@ -733,14 +784,14 @@ module Props =
         | ContentInset of Insets
         | Html of string
         | InjectedJavaScript of string
-        | OnError of Func<NavState, unit>
-        | OnLoad of Func<NavState, unit>
-        | OnLoadEnd of Func<NavState, unit>
-        | OnLoadStart of Func<NavState, unit>
-        | OnNavigationStateChange of Func<NavState, unit>
-        | OnShouldStartLoadWithRequest of Func<bool>
-        | RenderError of Func<React.ReactElement>
-        | RenderLoading of Func<React.ReactElement>
+        | OnError of (NavState -> unit)
+        | OnLoad of (NavState -> unit)
+        | OnLoadEnd of (NavState -> unit)
+        | OnLoadStart of (NavState -> unit)
+        | OnNavigationStateChange of (NavState -> unit)
+        | OnShouldStartLoadWithRequest of (obj -> bool)
+        | RenderError of (unit -> React.ReactElement)
+        | RenderLoading of (unit -> React.ReactElement)
         | ScrollEnabled of bool
         | StartInLoadingState of bool
         | Style of IStyle list
@@ -754,8 +805,8 @@ module Props =
     type SegmentedControlIOSProperties =
         | Enabled of bool
         | Momentary of bool
-        | OnChange of Func<NativeSyntheticEvent<NativeSegmentedControlIOSChangeEvent>, unit>
-        | OnValueChange of Func<string, unit>
+        | OnChange of (NativeSyntheticEvent<NativeSegmentedControlIOSChangeEvent> -> unit)
+        | OnValueChange of (string -> unit)
         | SelectedIndex of float
         | TintColor of string
         | Values of ResizeArray<string>
@@ -788,7 +839,7 @@ module Props =
             | Animating of bool
             | Color of string
             | HidesWhenStopped of bool
-            | OnLayout of Func<obj, unit>
+            | OnLayout of (obj -> unit)
             | Size of Size
             | Style of IStyle list
             | Ref of Ref<ActivityIndicatorIOS>
@@ -800,7 +851,7 @@ module Props =
         | MinimumDate of DateTime
         | MinuteInterval of float
         | Mode of DatePickerIOSMode
-        | OnDateChange of Func<DateTime, unit>
+        | OnDateChange of (DateTime -> unit)
         | TimeZoneOffsetInMinutes of float
         | Ref of Ref<DatePickerIOS>
         interface IDatePickerIOSProperties
@@ -811,27 +862,25 @@ module Props =
         | DrawerPosition of DrawerLayoutAndroidPosition
         | DrawerWidth of float
         | KeyboardDismissMode of KeyboardDismissMode
-        | OnDrawerClose of (unit->unit)
-        | OnDrawerOpen of (unit->unit)
-        | OnDrawerSlide of Func<DrawerSlideEvent, unit>
-        | OnDrawerStateChanged of Func<ScrollState, unit>
-        | RenderNavigationView of Func<obj>
+        | OnDrawerClose of (unit -> unit)
+        | OnDrawerOpen of (unit -> unit)
+        | OnDrawerSlide of (DrawerSlideEvent -> unit)
+        | OnDrawerStateChanged of (ScrollState -> unit)
+        | RenderNavigationView of (unit -> obj)
         | StatusBarBackgroundColor of obj
         | Ref of Ref<obj>
         interface IDrawerLayoutAndroidProperties
 
     module Picker =
         type PickerIOSItemProperties =
-            | Value of U2<string, float>
+            | Value of U2<string, int>
             | Label of string
-
 
         type PickerItemProperties =
             | Label of string // REQUIRED!
-            | Value of obj
+            | Value of string
             | Color of string
             | TestID of string
-
 
         type PickerPropertiesIOS =
             | ItemStyle of IStyle list
@@ -846,8 +895,8 @@ module Props =
             interface IPickerProperties
 
         type PickerProperties =
-            | OnValueChange of Func<obj, float, unit>
-            | SelectedValue of obj
+            | OnValueChange of (string -> int -> unit)
+            | SelectedValue of string
             | Style of IStyle list
             | TestId of string
             | Ref of Ref<Picker>
@@ -897,7 +946,7 @@ module Props =
         interface IRefreshControlProperties
 
     type RefreshControlProperties =
-        | OnRefresh of (unit->unit)
+        | OnRefresh of (unit -> unit)
         | Refreshing of bool
         | Ref of Ref<RefreshControl>
         interface IRefreshControlProperties
@@ -916,8 +965,8 @@ module Props =
         | Disabled of bool
         | MaximumValue of float
         | MinimumValue of float
-        | OnSlidingComplete of Func<float, unit>
-        | OnValueChange of Func<float, unit>
+        | OnSlidingComplete of (float -> unit)
+        | OnValueChange of (float -> unit)
         | Step of float
         | Style of IStyle list
         | TestID of string
@@ -931,23 +980,18 @@ module Props =
         | MinimumValue of float
         | MinimumTrackImage of obj
         | MinimumTrackTintColor of string
-        | OnSlidingComplete of (unit->unit)
-        | OnValueChange of Func<float, unit>
+        | OnSlidingComplete of (unit -> unit)
+        | OnValueChange of (float -> unit)
         | Step of float
         | Style of IStyle list
         | Value of float
         | Ref of Ref<SliderIOS>
         interface ISliderIOSProperties
 
-    type SwitchIOSStyle =
-        | Height of float
-        | Width of float
-        interface ISwitchIOSStyle
-
     type SwitchIOSProperties =
         | Disabled of bool
         | OnTintColor of string
-        | OnValueChange of Func<bool, unit>
+        | OnValueChange of (bool -> unit)
         | ThumbTintColor of string
         | TintColor of string
         | Value of bool
@@ -964,7 +1008,7 @@ module Props =
         | BorderRadius of float
         | BorderTopLeftRadius of float
         | BorderTopRightRadius of float
-        | Overflow of Overflow
+        | Overflow of ImageOverflow
         | OverlayColor of string
         | TintColor of string
         | Opacity of float
@@ -989,15 +1033,15 @@ module Props =
         | Accessible of bool
         | CapInsets of Insets
         | DefaultSource of IImageSourceProperties list
-        | OnError of Func<obj, unit>
-        | OnProgress of (unit->unit)
+        | OnError of (obj -> unit)
+        | OnProgress of (unit -> unit)
         interface IImagePropertiesIOS
 
     type ImageProperties =
-        | OnLayout of Func<LayoutChangeEvent, unit>
-        | OnLoad of (unit->unit)
-        | OnLoadEnd of (unit->unit)
-        | OnLoadStart of (unit->unit)
+        | OnLayout of (LayoutChangeEvent -> unit)
+        | OnLoad of (unit -> unit)
+        | OnLoadEnd of (unit -> unit)
+        | OnLoadStart of (unit -> unit)
         | ResizeMode of ResizeMode
         | Source of IImageSourceProperties list
         | Style of IStyle list
@@ -1013,8 +1057,8 @@ module Props =
         | Subtitle of string
         | HasLeftCallout of bool
         | HasRightCallout of bool
-        | OnLeftCalloutPress of (unit->unit)
-        | OnRightCalloutPress of (unit->unit)
+        | OnLeftCalloutPress of (unit -> unit)
+        | OnRightCalloutPress of (unit -> unit)
         | Id of string
 
     type MapViewRegion =
@@ -1046,9 +1090,9 @@ module Props =
         interface IMapViewPropertiesAndroid
 
     type MapViewProperties =
-        | OnAnnotationPress of (unit->unit)
-        | OnRegionChange of Func<MapViewRegion, unit>
-        | OnRegionChangeComplete of Func<MapViewRegion, unit>
+        | OnAnnotationPress of (unit -> unit)
+        | OnRegionChange of (MapViewRegion -> unit)
+        | OnRegionChangeComplete of (MapViewRegion -> unit)
         | PitchEnabled of bool
         | Region of MapViewRegion
         | RotateEnabled of bool
@@ -1064,8 +1108,20 @@ module Props =
         | AnimationType of AnimationType
         | Transparent of bool
         | Visible of bool
-        | OnRequestClose of (unit->unit)
-        | OnShow of Func<NativeSyntheticEvent<obj>, unit>
+        | OnRequestClose of (unit -> unit)
+        | OnShow of (NativeSyntheticEvent<obj> -> unit)
+
+    type IButtonProperties =
+        interface end
+
+    type ButtonProperties =
+        | Title of string
+        | OnPress of (unit -> unit)
+        | Disabled of bool
+        | Color of string
+        | TestID of string
+        | HasTVPreferredFocus of bool
+        interface IButtonProperties
 
     type ITouchableHighlightProperties =
         interface end
@@ -1104,19 +1160,19 @@ module Props =
         | DelayPressOut of float
         | Disabled of bool
         | HitSlop of obj
-        | OnLayout of Func<LayoutChangeEvent, unit>
-        | OnLongPress of (unit->unit)
-        | OnPress of (unit->unit)
-        | OnPressIn of (unit->unit)
-        | OnPressOut of (unit->unit)
+        | OnLayout of (LayoutChangeEvent -> unit)
+        | OnLongPress of (unit -> unit)
+        | OnPress of (unit -> unit)
+        | OnPressIn of (unit -> unit)
+        | OnPressOut of (unit -> unit)
         | Style of IStyle list
         | PressRetentionOffset of obj
         interface ITouchableWithoutFeedbackProperties
 
     type TouchableHighlightProperties =
         | ActiveOpacity of float
-        | OnHideUnderlay of (unit->unit)
-        | OnShowUnderlay of (unit->unit)
+        | OnHideUnderlay of (unit -> unit)
+        | OnShowUnderlay of (unit -> unit)
         | Style of IStyle list
         | UnderlayColor of string
         interface ITouchableHighlightProperties
@@ -1131,9 +1187,9 @@ module Props =
         interface ITouchableNativeFeedbackProperties
 
     type NavigationBarRouteMapper =
-        | Title of Func<Route, Navigator, float, NavState, React.ReactElement>
-        | LeftButton of Func<Route, Navigator, float, NavState, React.ReactElement>
-        | RightButton of Func<Route, Navigator, float, NavState, React.ReactElement>
+        | Title of (Route -> Navigator -> float -> NavState -> React.ReactElement)
+        | LeftButton of (Route -> Navigator -> float -> NavState -> React.ReactElement)
+        | RightButton of (Route -> Navigator -> float -> NavState -> React.ReactElement)
 
     type NavigationBarProperties =
         | Navigator of Navigator
@@ -1145,14 +1201,14 @@ module Props =
         interface end
 
     type NavigatorProperties =
-        | ConfigureScene of Func<Route, ResizeArray<Route>, SceneConfig>
+        | ConfigureScene of (Route -> ResizeArray<Route> -> SceneConfig)
         | InitialRoute of Route
         | InitialRouteStack of ResizeArray<Route>
         | NavigationBar of React.ReactElement // React.ReactElement option
         | Navigator of Navigator
-        | OnDidFocus of (unit->unit)
-        | OnWillFocus of (unit->unit)
-        | RenderScene of Func<Route, Navigator, React.ReactElement>
+        | OnDidFocus of (unit -> unit)
+        | OnWillFocus of (unit -> unit)
+        | RenderScene of (Route -> Navigator -> React.ReactElement)
         | SceneStyle of ViewStyle list
         | DebugOverlay of bool
         interface INavigatorProperties
@@ -1161,7 +1217,7 @@ module Props =
         type TabBarItemProperties =
             | Badge of U2<string, float>
             | Icon of U2<obj, string>
-            | OnPress of (unit->unit)
+            | OnPress of (unit -> unit)
             | Selected of bool
             | SelectedIcon of U2<obj, string>
             | Style of IStyle list
@@ -1225,8 +1281,8 @@ module Props =
         | IndicatorStyle of IndicatorStyle
         | MaximumZoomScale of float
         | MinimumZoomScale of float
-        | OnRefreshStart of (unit->unit)
-        | OnScrollAnimationEnd of (unit->unit)
+        | OnRefreshStart of (unit -> unit)
+        | OnScrollAnimationEnd of (unit -> unit)
         | ScrollEnabled of bool
         | ScrollEventThrottle of float
         | ScrollIndicatorInsets of Insets
@@ -1245,12 +1301,18 @@ module Props =
         | ScrollPerfTag of string
         interface IScrollViewPropertiesAndroid
 
+    [<RequireQualifiedAccess; StringEnum>]
+    type KeyboardShouldPersistTapsProperties =
+    | Never
+    | Always
+    | Handled
+
     type ScrollViewProperties =
         | ContentContainerStyle of ViewStyle list
         | Horizontal of bool
         | KeyboardDismissMode of string
-        | KeyboardShouldPersistTaps of bool
-        | OnScroll of Func<obj, unit>
+        | KeyboardShouldPersistTaps of KeyboardShouldPersistTapsProperties
+        | OnScroll of (obj -> unit)
         | PagingEnabled of bool
         | RemoveClippedSubviews of bool
         | ShowsHorizontalScrollIndicator of bool
@@ -1260,65 +1322,65 @@ module Props =
         | Ref of Ref<ScrollView>
         interface IScrollViewProperties
 
-    type ListViewProperties<'a> = 
+    type ListViewProperties<'a> =
         | DataSource of ListViewDataSource<'a>
         | EnableEmptySections of bool
         | InitialListSize of float
-        | OnChangeVisibleRows of Func<ResizeArray<obj>, ResizeArray<obj>, unit>
-        | OnEndReached of (unit->unit)
+        | OnChangeVisibleRows of (ResizeArray<obj> -> ResizeArray<obj> -> unit)
+        | OnEndReached of (unit -> unit)
         | OnEndReachedThreshold of float
         | PageSize of float
         | RemoveClippedSubviews of bool
-        | RenderFooter of Func<React.ReactElement>
-        | RenderHeader of Func<React.ReactElement>
-        | RenderRow of Func<'a, U2<string, float>, U2<string, float>, bool, React.ReactElement>
-        | RenderScrollComponent of Func<ScrollViewProperties, React.ReactElement>
-        | RenderSectionHeader of Func<obj, U2<string, float>, React.ReactElement>
-        | RenderSeparator of Func<U2<string, float>, U2<string, float>, bool, React.ReactElement>
+        | RenderFooter of (unit -> React.ReactElement)
+        | RenderHeader of (unit -> React.ReactElement)
+        | RenderRow of ('a -> U2<string, float> -> U2<string, float> -> bool -> React.ReactElement)
+        | RenderScrollComponent of (ScrollViewProperties -> React.ReactElement)
+        | RenderSectionHeader of (obj -> U2<string, float> -> React.ReactElement)
+        | RenderSeparator of (U2<string, float> -> U2<string, float> -> bool -> React.ReactElement)
         | ScrollRenderAheadDistance of float
         | Ref of Ref<obj>
         interface IListViewProperties
 
-    type FlatListRenderItemSeparator = { highlight : Func<unit, unit>; unhighlight : Func<unit, unit> }
+    type FlatListRenderItemSeparator = { highlight : (unit -> unit); unhighlight : (unit -> unit) }
     type FlatListRenderItemInfo<'a> = { item : 'a; index : float; separators : FlatListRenderItemSeparator }
 
     type GetItemLayoutResult = { length : float; offset : float; index : float }
-    
+
     type ViewToken<'a> = { item : 'a; key : string; index : float; isViewable : bool; section : obj }
     type OnViewableItemsChangedInfo<'a> = { viewableItems : ViewToken<'a> []; changed : ViewToken<'a> [] }
 
     type ViewabilityConfig = { minimumViewTime : float; viewAreaCoveragePercentThreshold : float; itemVisiblePercentThreshold : float; waitForInteraction : bool }
 
     type FlatListProperties<'a> =
-        | ItemSeparatorComponent of React.ReactElement
-        | ListEmptyComponent of React.ReactElement
-        | ListFooterComponent of React.ReactElement
-        | ListHeaderComponent of React.ReactElement
+        | ItemSeparatorComponent of (unit -> React.ReactElement)
+        | ListEmptyComponent of (unit -> React.ReactElement)
+        | ListFooterComponent of (unit -> React.ReactElement)
+        | ListHeaderComponent of (unit -> React.ReactElement)
         | ColumnWrapperStyle of IStyle list
         | ExtraData of obj
-        | GetItemLayout of Func<ResizeArray<'a>, GetItemLayoutResult>
+        | GetItemLayout of (ResizeArray<'a> -> GetItemLayoutResult)
         | Horizontal of bool
-        | InitialNumToRender of float
+        | InitialNumToRender of int
         | InitialScrollIndex of float
-        | KeyExtractor of Func<'a, float, string>
+        | KeyExtractor of ('a -> int -> string)
         | LegacyImplementation of bool
-        | NumColumns of float
-        | OnEndReached of Func<float, unit>
+        | NumColumns of int
+        | OnEndReached of (float -> unit)
         | OnEndReachedThreshold of float
-        | OnRefresh of Func<unit, unit>
-        | OnViewableItemsChanged of Func<OnViewableItemsChangedInfo<'a>, unit>
+        | OnRefresh of (unit -> unit)
+        | OnViewableItemsChanged of (OnViewableItemsChangedInfo<'a> -> unit)
         | Refreshing of bool
         | RemoveClippedSubviews of bool
-        | RenderItem of Func<FlatListRenderItemInfo<'a>, React.ReactElement>
+        | RenderItem of (FlatListRenderItemInfo<'a> -> React.ReactElement)
         | ViewabilityConfig of ViewabilityConfig
         | Ref of Ref<obj>
-        interface IFlatListProperties
+        interface IFlatListProperties<'a>
 
     type SwipeableListViewProps<'a> =
         | DataSource of SwipeableListViewDataSource<'a> // REQUIRED!
         | MaxSwipeDistance of float
-        | RenderRow of Func<'a, U2<string, float>, U2<string, float>, bool, React.ReactElement> // REQUIRED!
-        | RenderQuickActions of Func<'a, string, string, React.ReactElement> // REQUIRED!
+        | RenderRow of ('a -> U2<string, float> -> U2<string, float> -> bool -> React.ReactElement) // REQUIRED!
+        | RenderQuickActions of ('a -> string -> string -> React.ReactElement) // REQUIRED!
 
     type ActionSheetIOSOptions =
         | Title of string
@@ -1338,18 +1400,18 @@ module Props =
         | MaxDate of U2<DateTime, float>
 
     type PanResponderCallbacks =
-        | OnMoveShouldSetPanResponder of Func<GestureResponderEvent, PanResponderGestureState, bool>
-        | OnStartShouldSetPanResponder of Func<GestureResponderEvent, PanResponderGestureState, unit>
-        | OnPanResponderGrant of Func<GestureResponderEvent, PanResponderGestureState, unit>
-        | OnPanResponderMove of Func<GestureResponderEvent, PanResponderGestureState, unit>
-        | OnPanResponderRelease of Func<GestureResponderEvent, PanResponderGestureState, unit>
-        | OnPanResponderTerminate of Func<GestureResponderEvent, PanResponderGestureState, unit>
-        | OnMoveShouldSetPanResponderCapture of Func<GestureResponderEvent, PanResponderGestureState, bool>
-        | OnStartShouldSetPanResponderCapture of Func<GestureResponderEvent, PanResponderGestureState, bool>
-        | OnPanResponderReject of Func<GestureResponderEvent, PanResponderGestureState, unit>
-        | OnPanResponderStart of Func<GestureResponderEvent, PanResponderGestureState, unit>
-        | OnPanResponderEnd of Func<GestureResponderEvent, PanResponderGestureState, unit>
-        | OnPanResponderTerminationRequest of Func<GestureResponderEvent, PanResponderGestureState, bool>
+        | OnMoveShouldSetPanResponder of (GestureResponderEvent -> PanResponderGestureState -> bool)
+        | OnStartShouldSetPanResponder of (GestureResponderEvent -> PanResponderGestureState -> unit)
+        | OnPanResponderGrant of (GestureResponderEvent -> PanResponderGestureState -> unit)
+        | OnPanResponderMove of (GestureResponderEvent -> PanResponderGestureState -> unit)
+        | OnPanResponderRelease of (GestureResponderEvent -> PanResponderGestureState -> unit)
+        | OnPanResponderTerminate of (GestureResponderEvent -> PanResponderGestureState -> unit)
+        | OnMoveShouldSetPanResponderCapture of (GestureResponderEvent -> PanResponderGestureState -> bool)
+        | OnStartShouldSetPanResponderCapture of (GestureResponderEvent -> PanResponderGestureState -> bool)
+        | OnPanResponderReject of (GestureResponderEvent -> PanResponderGestureState -> unit)
+        | OnPanResponderStart of (GestureResponderEvent -> PanResponderGestureState -> unit)
+        | OnPanResponderEnd of (GestureResponderEvent -> PanResponderGestureState -> unit)
+        | OnPanResponderTerminationRequest of (GestureResponderEvent -> PanResponderGestureState -> bool)
 
     module StatusBar =
         type StatusBarPropertiesIOS =
@@ -1427,21 +1489,30 @@ module Props =
 open Props
 module R = Fable.Helpers.React
 
+[<Emit("$0")>]
+// density independent pixels
+let Dip (_: float): ISizeUnit = jsNative
+
+[<Emit("$0 + \"%\"")>]
+// percents
+let Pct (_: float): ISizeUnit = jsNative
+
 [<Emit("require($0)")>]
 // Use `require` to load a local image
 let inline localImage (path:string) : IImageSourceProperties list = jsNative
 
 let inline createElement(c: React.ComponentClass<'T>, props: 'P list, children: React.ReactElement list) =
-    applySpread R.createEl (c, keyValueList CaseRules.LowerFirst props, children)
+    R.createElement (c, keyValueList CaseRules.LowerFirst props, children)
 
 let inline internal createElementWithObjProps(c: React.ComponentClass<'T>, props: obj, children: React.ReactElement list) =
-    applySpread R.createEl (c, props, children)
+    R.createElement (c, props, children)
 
 let inline text (props:TextProperties list) (text:string): React.ReactElement =
     createElement(RN.Text, props, [React.str text])
 
 let inline textInput (props: ITextInputProperties list) (text:string): React.ReactElement =
-    createElement(RN.TextInput, props, [R.str text])
+    let valueProp = TextInput.TextInputProperties.Value text :> ITextInputProperties
+    createElement(RN.TextInput, valueProp :: props, [])
 
 let inline createToolbarAction(title:string,showStatus:ToolbarActionShowStatus) : ToolbarAndroidAction =
     createObj [
@@ -1578,12 +1649,24 @@ let inline listView<'a> (dataSource:ListViewDataSource<'a>) (props: IListViewPro
             createObj ["dataSource" ==> dataSource],
             keyValueList CaseRules.LowerFirst props), [])
 
-let inline flatList<'a> (data:'a []) (props: IFlatListProperties list)  : React.ReactElement =
+let inline flatList<'a> (data:'a []) (props: FlatListProperties<'a> list)  : React.ReactElement =
+    // Some of FlatList properties are upper case:
+    // https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/types/react-native/index.d.ts#L3608-L3623
+    let pascalCaseProps, camelCaseProps =
+      List.partition (function
+                      | ItemSeparatorComponent _ -> true
+                      | ListEmptyComponent _ -> true
+                      | ListFooterComponent _ -> true
+                      | ListHeaderComponent _ -> true
+                      | _ -> false)
+                      props
+
     createElementWithObjProps(
       RN.FlatList,
       !!JS.Object.assign(
             createObj ["data" ==> data],
-            keyValueList CaseRules.LowerFirst props), [])
+            keyValueList CaseRules.LowerFirst camelCaseProps,
+            keyValueList CaseRules.None pascalCaseProps), [])
 
 let inline mapView (props:IMapViewProperties list) (children: React.ReactElement list): React.ReactElement =
     createElement(
@@ -1591,10 +1674,17 @@ let inline mapView (props:IMapViewProperties list) (children: React.ReactElement
       props,
       children)
 
-let inline modal (props:ModalProperties list) : React.ReactElement =
+let inline modal (props:ModalProperties list) (children: React.ReactElement list): React.ReactElement =
     createElement(
       RN.Modal,
-      props, [])
+      props,
+      children)
+
+let inline button (props:IButtonProperties list) (children: React.ReactElement list) : React.ReactElement =
+    createElement(
+      RN.Button,
+      props,
+      children)
 
 let inline touchableWithoutFeedback (props:ITouchableWithoutFeedbackProperties list) (children: React.ReactElement list): React.ReactElement =
     createElement(
@@ -1635,11 +1725,6 @@ let inline viewPagerAndroid (props: IViewPagerAndroidProperties list) (children:
 let inline navigator (props:INavigatorProperties list) : React.ReactElement =
     createElement(
       RN.Navigator,
-      props, [])
-
-let inline styleSheet (props:StyleSheetProperties list) : React.ReactElement =
-    createElement(
-      RN.StyleSheet,
       props, [])
 
 let inline tabBarItem (props:ITabBarItemProperties list) : React.ReactElement =
@@ -1729,12 +1814,6 @@ let inline newDataSource<'a> (elements:'a []) =
 let inline updateDataSource<'a> (data:'a []) (dataSource : ListViewDataSource<'a>) : ListViewDataSource<'a> =
     dataSource.cloneWithRows(!!data)
 
-// [<Emit(typeof<React.Emitter>, "Com")>]
-// let createComponent<'T,'P,'S when 'T :> React.Component<'P,'S>> (props: 'P) (children: React.ReactElement list): React.ReactElement = jsNative
-
-// [<Emit(typeof<React.Emitter>, "Com")>]
-// let createScene<'T,'P,'S when 'T :> React.Component<'P,'S>> (props: 'P) : React.ReactElement = jsNative
-
 [<Import("Buffer","buffer")>]
 [<Emit("$0.from($1).toString($2)")>]
 let encode (text: string, encoding:string) : string = jsNative
@@ -1742,7 +1821,7 @@ let encode (text: string, encoding:string) : string = jsNative
 let encodeBase64 (text: string) : string = encode(text,"base64")
 let encodeAscii (text: string) : string = encode(text,"ascii")
 
-    
+
 [<Import("BackHandler","react-native")>]
 let private BackHandler = obj()
 
@@ -1819,6 +1898,64 @@ module ImageStore =
             ImageStore?addImageFromBase64(imageData, onSuccess, onError) |> ignore
         )
 
+/// ImageEditor contains functions which help to deal with image data.
+module ImageEditor =
+    [<RequireQualifiedAccess>]
+    type ResizeMode =
+    | Contain
+    | Cover
+    | Stretch
+
+    type CropData() =
+        let data = createObj [  ]
+
+        member this.SetSize(width:int,height:int) =
+            let size =
+                createObj
+                    [ "width" ==> width
+                      "height" ==> height ]
+
+            data?size <- size
+            this
+
+        member this.SetDisplaySize(width:int,height:int) =
+            let size =
+                createObj
+                    [ "width" ==> width
+                      "height" ==> height ]
+
+            data?displaySize <- size
+            this
+
+        member this.SetOffset(x:int,y:int) =
+            let offset =
+                createObj
+                    [ "x" ==> x
+                      "y" ==> y ]
+
+            data?offset <- offset
+            this
+
+        member this.SetResizeMode(mode:ResizeMode) =
+
+            data?displaysize <-
+                match mode with
+                | ResizeMode.Contain -> "contain"
+                | ResizeMode.Cover -> "cover"
+                | ResizeMode.Stretch -> "stretch"
+            this
+
+
+    [<Import("ImageEditor","react-native")>]
+    let private ImageEditor = obj()
+
+    /// Crop the image specified by the URI param. If URI points to a remote image, it will be downloaded automatically.
+    /// If the image cannot be loaded/downloaded, the failure callback will be called.
+    let cropImage (uri:string) (cropData:CropData) : JS.Promise<string> =
+        Promise.create(fun onSuccess onError ->
+            ImageEditor?cropImage(uri, cropData?data, onSuccess, onError) |> ignore
+        )
+
 module Toast =
     [<Import("ToastAndroid","react-native")>]
     let private Toast = obj()
@@ -1858,3 +1995,14 @@ module Storage =
     let [<PassGenerics>] save<'a> (k:string) (v:'a): JS.Promise<unit> =
         !!Globals.AsyncStorage.setItem(k, toJson v)
 
+module Platform =
+    type OS<'a> =
+        | Ios of 'a
+        | Android of 'a
+        | Macos of 'a
+        | Windows of 'a
+        | Web of 'a
+        | Default of 'a
+
+    let select<'a> (specifics: OS<'a> list): 'a =
+        Globals.Platform.select (keyValueList CaseRules.LowerFirst specifics)