+++ src/extra/react-native-image-resizer/Fable.Helpers.ReactNativeImageResizer.fs
@@ -6,9 +6,13 @@ open Fable.Import.ReactNativeImageResizer
 
 open Fable.Core.JsInterop
 
-
 /// A React Native module that can create scaled versions of local images (also supports the assets library on iOS).
 module ImageResizer =
+    type ResizeResult =
+       abstract uri: string
+       abstract path: string
+       abstract name: string
+       abstract size: int
 
     /// <summary>Creates a scaled version of a local images.</summary>
     /// <param name="path">Path of image file, or a base64 encoded image string prefixed with 'data:image/imagetype' where imagetype is jpeg or png.</param>
@@ -16,5 +20,6 @@ module ImageResizer =
     /// <param name="maxHeight">Image max height (ratio is preserved)</param>
     /// <param name="compressFormat">Can be either JPEG, PNG or WEBP (Android only).</param>
     /// <param name="quality">A number between 0 and 100. Used for the JPEG compression.</param>
-    let createResizedImage (path: string, maxWidth: int, maxHeight: int, compressFormat: string, quality: int) : JS.Promise<string> =
-        ReactNativeImageResizer.Globals.ImageResizer?createResizedImage(path, maxWidth, maxHeight, compressFormat, quality) |> unbox
+    let createResizedImage (path: string, maxWidth: int, maxHeight: int, compressFormat: string, quality: int) : JS.Promise<ResizeResult> =
+        ReactNativeImageResizer.Globals.ImageResizer?createResizedImage(path, maxWidth, maxHeight, compressFormat, quality)
+        |> unbox