+++ RELEASE_NOTES.md
@@ -1,11 +1,52 @@
-### 1.3.0-alpha003
-* Fix flat list
+### 2.0.0-alpha-001
+* Fixes by @iyegoroff: PR #23
 
-### 1.3.0-alpha002
-* Fix flat list
+### 1.7.2
+* Some of FlatList properties are upper case: @iyegoroff
 
-### 1.3.0-alpha001
-* RN 0.47 support
+### 1.7.1
+* KeyboardShouldPersistTaps changed for ScrollViews
+
+### 1.7.0
+* async-safe, static lifecycle hooks - https://fb.me/react-async-component-lifecycle-hooks
+
+### 1.6.6
+* KeyboardShouldPersistTaps changed for ScrollViews
+
+### 1.6.5
+* FlatList's KeyExtractor by Index is an int (@forki)
+
+### 1.6.3
+* Add a button (#20 @forki)
+
+### 1.6.2
+* Adding react-native-device-info helper (#19 @forki)
+
+### 1.6.1
+* Fix #17: textInput on IOS (@ddunlea)
+
+### 1.6.0
+* Fix compat issues with Fable update
+
+### 1.5.2
+* Type info for items in FlatList
+
+### 1.5.0
+* Barcode scanning via react-native-camera
+* Picker values as string
+
+### 1.4.0
+* Add ReactNativeSignatureView support to the project
+* Add react-native-modal-datetime-picker bindings - https://github.com/fable-compiler/fable-react-native/pull/11
+
+### 1.3.0
+* BREAKING: RN 0.47 support - this is breaking since React Native is breaking in some cases
+* BREAKING: ImageResizer fixes for v1.0 - https://github.com/fable-compiler/fable-react-native/pull/10
+* Add ImageEditor bindings - https://github.com/fable-compiler/fable-react-native/pull/6
+* Add ReactNativeSignatureView support - https://github.com/fable-compiler/fable-react-native/pull/7
+* Async launchCamera and launchImageLibrary functions - https://github.com/fable-compiler/fable-react-native/pull/8
+* Add AspectRatio to FlexStyle
+* Flat list support
 
 ### 1.2.3
 * SignaturePad